move enterprise/internal/codycontext to internal/codycontext
