+++ enterprise/cmd/frontend/internal/context/BUILD.bazel
@@ -8,9 +8,9 @@ go_library(
     deps = [
         "//cmd/frontend/enterprise",
         "//enterprise/cmd/frontend/internal/context/resolvers",
-        "//enterprise/internal/codycontext:context",
         "//enterprise/internal/database",
         "//internal/codeintel",
+        "//internal/codycontext:context",
         "//internal/conf/conftypes",
         "//internal/database",
         "//internal/embeddings",
